Refactor formatters para evitar imports circulares en py2
