+++ pydatajson/response_formatters/validation_response_formatter.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+
+from __future__ import unicode_literals
+import abc
+
+
+class ValidationResponseFormatter(object):
+
+    def __init__(self, response):
+        self.response = response
+
+    @abc.abstractmethod
+    def format(self):
+        raise NotImplementedError
